+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Release 1.2.0 (2016-02-01)
+
+- Added quote and squote
+- Added b32enc and b32dec
+- add now takes varargs
+- biggest now takes varargs
+
 # Release 1.1.0 (2015-12-29)
 
 - Added #4: Added contains function. strings.Contains, but with the arguments